ml-api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
joblib==1.4.2
scikit-learn==1.5.0
pandas==2.2.2
xgboost==2.0.3
numpy==1.26.4
tensorflow==2.16.1
pydantic==2.7.1
psycopg2-binary==2.9.9